refactor(cli): extract output.ts and help.ts from qmd.ts (batch 2)
- output.ts: search result rendering for all 6 formats (outputResults,
  printEmptySearchResults) plus editor-link helpers (buildEditorUri,
  termLink, getEditorUriTemplate). Not a pure formatter — it reads the
  active index and resolves --full-path targets, so it imports
  context.ts and commands/docs.ts; direction stays one-way.
- help.ts: showHelp/showSkillsHelp/printDoctorHint/showVersion/
  readPackageJson. readPackageJson keeps working because help.ts sits
  at the same depth as qmd.ts (../../package.json).
- commands/docs.ts created early with renderFullPath (spec §8.3 a1):
  outputResults needs it and docs.ts must not import output.ts.
  computeDisplayPath moves along as-is; it is pre-existing dead code
  (no callers in src/ or test/), kept per pure-move discipline.
- qmd.ts re-exports buildEditorUri/termLink (I4).

Verified: tsc clean, build ok, 32 snapshots byte-identical,
cli/lazy-llm-import/exit-lifecycle green, cli.test.ts 146/146.
